set dtype

+++ sklearnex/utils/tests/test_finite.py
@@ -37,7 +37,7 @@ from sklearnex.utils import _assert_all_finite
 )
 @pytest.mark.parametrize("allow_nan", [False, True])
 def test_sum_infinite_actually_finite(dtype, shape, allow_nan):
-    X = np.array(shape)
+    X = np.array(shape, dtype=dtype)
     X.fill(np.finfo(dtype).max)
     _assert_all_finite(X, allow_nan=allow_nan)
